Crate linear_assignment

This is a library for solving the linear assignment problem. For this problem, it is helpful to think of the data as a bipartite graph, with weighted, undirected edges fully connecting all elements of set U to all elements of set V.

Structs

MatrixSize

Traits

Weight

Functions

solver
Find a solution to the assignment problem in matrix.

Type Definitions

Edge
An edge between U and V: (u, v)